% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DESeqResMeta-class}
\alias{DESeqResMeta-class}
\title{An S4 class extending the DESeqResults object containing additional fields for differential expression data.}
\description{
This class is used to separate and identify relevant pieces of data by various visualizations in this package.
}
\section{Slots}{

\describe{
\item{\code{numDE}}{The total number of differentially expressed genes in the result set.}

\item{\code{numUp}}{The number of upregulated differentially expressed genes in the result set.}

\item{\code{numDown}}{The number of downregulated differentially expressed genes in the result set.}

\item{\code{allNames}}{The names of all differentially expressed genes in the result set.}

\item{\code{upNames}}{The names of all upregulated differentially expressed genes in the result set.}

\item{\code{downNames}}{The names of all downregulated differentially expressed genes in the result set.}

\item{\code{case}}{The "case" condition that was used to make the contrast for this result set.}

\item{\code{control}}{The "control" condition that was used to make the contrast for this result set.}

\item{\code{contrast}}{The contrast that was made for this result set.  I.E. case_vs_control}

\item{\code{design_field}}{The field in the provided target data that was used for experimental design in differential expression.
I.E. Condition_Time.}
}}

