% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{set_output_mode}
\alias{set_output_mode}
\title{Determine if visualizations are written to file, printed to screen, or both.}
\usage{
set_output_mode(toggle = "both")
}
\arguments{
\item{toggle}{Desired output mode. Determines how output plot data will be displayed. Options are "screen", "file", or "both".}
}
\value{
This function does not return a value.
}
\description{
This function initializes the output mode for the package to determine
if visualizations should be written to file, printed to screen, or both.
}
\examples{
\dontrun{

#Don't save files, just print to the screen.
set_output_mode("screen")

#Print to the screen and save files to appropriate locations.
set_output_mode("both")

#Save files to appropriate locations, but do not print them to screen.
set_output_mode("file")

}
}
\keyword{display}
\keyword{file}
\keyword{output}
